+++ src/legendre/mod.rs
@@ -5,6 +5,8 @@
 
 //! Elliptic integral functions in Legendre's form.
 
+mod ellipd;
+mod ellipdinc;
 mod ellipe;
 mod ellipeinc;
 mod ellipf;
@@ -12,6 +14,8 @@ mod ellipk;
 mod ellippi;
 mod ellippiinc;
 
+pub use ellipd::ellipd;
+pub use ellipdinc::ellipdinc;
 pub use ellipe::ellipe;
 pub use ellipeinc::ellipeinc;
 pub use ellipf::ellipf;